server/PathonDB.DatabaseClient/Middlewares/Select/HasValidTable.cs
using System.Linq;
using PathonDB.DatabaseClient.Exceptions.General;
using PathonDB.DatabaseClient.Middlewares.General;
using PathonDB.DatabaseClient.Models.Database;
using PathonDB.DatabaseClient.Utils;

namespace PathonDB.DatabaseClient.Middlewares.Select {
    public class HasValidTable : Middleware {
        private readonly IDatabase _database;

        public HasValidTable(IDatabase database) {
            _database = database;
        }

        public override bool Check(string query) {
            var tableName = SelectUtils.GetTableNameFromArguments(SelectUtils.GetArgumentsFromQuery(query));
            if(tableName == null || !_database.GetTableNames().Contains(tableName.ToLower())) throw new TableNotFoundException(tableName);

            return this.CheckNext(query);
        }
    }
}